big refactor. introduce vector to simplify movement. add a simple enemy.

+++ AngleComponent/src/main/java/io/asteroidsfx/anglecomponent/AngleComponent.java
@@ -4,6 +4,6 @@ import io.asteroidsfx.common.Component;
 
 public class AngleComponent extends Component {
 
-    public double angle;
+    public float angle;
 
 }
