+++ CircleColliderComponent/src/main/java/io/asteroidsfx/circlecollidercomponent/CircleColliderComponent.java
@@ -4,6 +4,6 @@ import io.asteroidsfx.common.Component;
 
 public class CircleColliderComponent extends Component {
 
-    public double radius;
+    public float radius;
 
 }
