+++ OutOfBoundsComponent/src/main/java/io/asteroidsfx/outofboundscomponent/OutOfBoundsComponent.java
@@ -4,8 +4,8 @@ import io.asteroidsfx.common.Component;
 
 public class OutOfBoundsComponent extends Component {
     public BoundsAction boundsAction;
-    public double leftExtent;
-    public double rightExtent;
-    public double topExtent;
-    public double bottomExtent;
+    public int leftExtent;
+    public int rightExtent;
+    public int topExtent;
+    public int bottomExtent;
 }
