+++ ShootComponent/src/main/java/io/asteroidsfx/shootcomponent/ShootComponent.java
@@ -5,7 +5,7 @@ import io.asteroidsfx.common.Component;
 public class ShootComponent extends Component {
 
     public boolean shootRequested = false;
-    public double roundsPerSecond;
-    public double velocity;
+    public float roundsPerSecond;
+    public float velocity;
     public long msLastFired = java.lang.System.currentTimeMillis();
 }
